+++ gpaw/directmin/etdm.py
@@ -30,6 +30,7 @@ class ETDM:
     def __init__(self,
                  searchdir_algo='l-bfgs-p',
                  linesearch_algo='swc-awc',
+                 partial_diagonalizer='Davidson',
                  update_ref_orbs_counter=20,
                  update_ref_orbs_canonical=False,
                  update_precond_counter=1000,
@@ -57,6 +58,8 @@ class ETDM:
         :param searchdir_algo: algorithm for calculating the search direction
         (e.g.LBFGS)
         :param linesearch_algo: line search (e.g. strong Wolfe conditions)
+        :param partial_diagonalizer: Algorithm to use for partial
+        diagonalization of the electronic Hessian if DO-GMF is used
         :param update_ref_orbs_counter: When to update C_ref
         :param update_ref_orbs_canonical: update C_ref to canonical orbitals
         :param update_precond_counter: when to update the preconditioner
@@ -108,9 +111,18 @@ class ETDM:
         self.orthonormalization = orthonormalization
         self.constraints = constraints
 
-        self.searchdir_algo = search_direction(searchdir_algo)
-        if self.searchdir_algo.name == 'l-bfgs-p' and not self.use_prec:
+        self.gmf = False
+        self.searchdir_algo = search_direction(
+            searchdir_algo, self, partial_diagonalizer)
+        sd_name = self.searchdir_algo.name.replace('-', '').lower().split('_')
+        if sd_name[0] == 'lbfgsp' and not self.use_prec:
             raise ValueError('Use l-bfgs-p with use_prec=True')
+        if len(sd_name) == 2:
+            if sd_name[1] == 'gmf':
+                self.searchdir_algo.name = sd_name[0]
+                self.gmf = True
+                self.g_vec_u_original = None
+                self.pd = partial_diagonalizer
 
         self.line_search = line_search_algorithm(linesearch_algo,
                                                  self.evaluate_phi_and_der_phi,
@@ -149,27 +161,38 @@ class ETDM:
 
     def __repr__(self):
 
-        sda_name = self.searchdir_algo.name
-        lsa_name = self.line_search.name
+        sda_name = self.searchdir_algo.name.replace('-', '').lower()
+        lsa_name = self.line_search.name.replace('-', '').lower()
+
+        add = ''
+        pd_add = ''
+        if self.gmf:
+            add = ' with minimum mode following'
+            pardi = {'Davidson': 'Finite difference generalized Davidson '
+                     'algorithm'}
+            pd_add = '       ' \
+                     'Partial diagonalizer: {}\n'.format(
+                         pardi[self.pd['name']])
 
         sds = {'sd': 'Steepest Descent',
-               'fr-cg': 'Fletcher-Reeves conj. grad. method',
-               'quick-min': 'Molecular-dynamics based algorithm',
-               'l-bfgs': 'L-BFGS algorithm',
-               'l-bfgs-p': 'L-BFGS algorithm with preconditioning',
-               'l-sr1p': 'Limited-memory SR1P algorithm'}
+               'frcg': 'Fletcher-Reeves conj. grad. method',
+               'quickmin': 'Molecular-dynamics based algorithm',
+               'lbfgs': 'L-BFGS algorithm',
+               'lbfgsp': 'L-BFGS algorithm with preconditioning',
+               'lsr1p': 'Limited-memory SR1P algorithm'}
 
-        lss = {'max-step': 'step size equals one',
+        lss = {'maxstep': 'step size equals one',
                'parabola': 'Parabolic line search',
-               'swc-awc': 'Inexact line search based on cubic interpolation,\n'
+               'swcawc': 'Inexact line search based on cubic interpolation,\n'
                           '                    strong and approximate Wolfe '
                           'conditions'}
 
-        repr_string = 'Direct minimisation using exponential ' \
+        repr_string = 'Direct minimisation' + add + ' using exponential ' \
                       'transformation.\n'
         repr_string += '       ' \
                        'Search ' \
-                       'direction: {}\n'.format(sds[sda_name])
+                       'direction: {}\n'.format(sds[sda_name] + add)
+        repr_string += pd_add
         repr_string += '       ' \
                        'Line ' \
                        'search: {}\n'.format(lss[lsa_name])
@@ -347,8 +370,6 @@ class ETDM:
         """
         with wfs.timer('Direct Minimisation step'):
             self.update_ref_orbitals(wfs, ham, dens)
-            with wfs.timer('Preconditioning:'):
-                precond = self.get_preconditioning(wfs, self.use_prec)
 
             a_vec_u = self.a_vec_u
             n_dim = self.n_dim
@@ -362,7 +383,19 @@ class ETDM:
                     self.get_energy_and_gradients(a_vec_u, n_dim, ham, wfs,
                                                   dens, c_ref)
             else:
-                g_vec_u = self.g_vec_u
+                g_vec_u = self.g_vec_u_original if self.gmf else self.g_vec_u
+
+            make_pd = False
+            if self.gmf:
+                with wfs.timer('Partial Hessian diagonalization'):
+                    self.searchdir_algo.update_eigenpairs(
+                        g_vec_u, wfs, ham, dens)
+                # The diagonal Hessian approximation must be positive-definite
+                make_pd = True
+
+            with wfs.timer('Preconditioning:'):
+                precond = self.get_preconditioning(
+                    wfs, self.use_prec, make_pd=make_pd)
 
             with wfs.timer('Get Search Direction'):
                 # calculate search direction according to chosen
@@ -494,6 +527,12 @@ class ETDM:
             phi, g_vec_u = self.get_energy_and_gradients(x_mat_u, n_dim,
                                                          ham, wfs, dens, c_ref)
 
+            # If GMF is used save the original gradient and invert the parallel
+            # projection onto the eigenvectors with negative eigenvalues
+            if self.gmf:
+                self.g_vec_u_original = deepcopy(g_vec_u)
+                g_vec_u = self.searchdir_algo.invert_parallel_grad(g_vec_u)
+
         der_phi = 0.0
         for k in p_mat_u:
             der_phi += g_vec_u[k].conj() @ p_mat_u[k]
@@ -532,12 +571,12 @@ class ETDM:
             # Erase memory of search direction algorithm
             self.searchdir_algo.reset()
 
-    def get_preconditioning(self, wfs, use_prec):
+    def get_preconditioning(self, wfs, use_prec, make_pd=False):
 
         if not use_prec:
             return None
 
-        if self.searchdir_algo.name == 'l-bfgs-p':
+        if self.searchdir_algo.name == 'lbfgsp':
             beta0 = self.searchdir_algo.beta_0
             gamma = 0.25
         else:
@@ -551,10 +590,16 @@ class ETDM:
             w = kpt.weight / (3.0 - wfs.nspins)
             if self.iters % counter == 0 or self.iters == 1:
                 self.hess[k] = self.get_hessian(kpt)
+                if make_pd:
+                    if self.dtype == float:
+                        self.hess[k] = np.abs(self.hess[k])
+                    else:
+                        self.hess[k] = np.abs(self.hess[k].real) \
+                            + 1.0j * np.abs(self.hess[k].imag)
             hess = self.hess[k]
             precond[k] = np.zeros_like(hess)
             correction = w * gamma * beta0 ** (-1)
-            if self.searchdir_algo.name != 'l-bfgs-p':
+            if self.searchdir_algo.name != 'lbfgsp':
                 correction = np.zeros_like(hess)
                 zeros = abs(hess) < 1.0e-4
                 correction[zeros] = 1.0
@@ -700,19 +745,23 @@ class ETDM:
 
     def todict(self):
 
-        return {'name': self.name,
-                'searchdir_algo': self.searchdir_algo.todict(),
-                'linesearch_algo': self.line_search.todict(),
-                'localizationtype': self.localizationtype,
-                'update_ref_orbs_counter': self.update_ref_orbs_counter,
-                'update_precond_counter': self.update_precond_counter,
-                'use_prec': self.use_prec,
-                'matrix_exp': self.matrix_exp,
-                'representation': self.representation,
-                'functional': self.func.todict(),
-                'orthonormalization': self.orthonormalization,
-                'constraints': self.constraints
-                }
+        ret = {'name': self.name,
+               'searchdir_algo': self.searchdir_algo.todict(),
+               'linesearch_algo': self.line_search.todict(),
+               'localizationtype': self.localizationtype,
+               'update_ref_orbs_counter': self.update_ref_orbs_counter,
+               'update_precond_counter': self.update_precond_counter,
+               'use_prec': self.use_prec,
+               'matrix_exp': self.matrix_exp,
+               'representation': self.representation,
+               'functional': self.func.todict(),
+               'orthonormalization': self.orthonormalization,
+               'constraints': self.constraints
+               }
+        if self.gmf:
+            ret['partial_diagonalizer'] = \
+                self.searchdir_algo.partial_diagonalizer.todict()
+        return ret
 
     def rotate_wavefunctions(self, wfs, a_vec_u, n_dim, c_ref):