+++ doc/releasenotes.rst
@@ -69,6 +69,9 @@ Git master branch
   of excited electronic states now includes :ref:`constrained optimization <mom>`
   useful for challenging charge transfer excited states.
 
+* The :ref:`direct optimization generalized mode following method <do-gmf>` (DO-GMF) for
+  variational calculations of excited electronic states was added.
+
 
 Version 22.8.0
 ==============